递送箱详情界面录音列表播放功能实现

+++ horizontalprogressbar/src/main/java/com/larunda/horizontalprogressbar/HorizontalProgressBarWithNunber.java
@@ -165,7 +165,7 @@ public class HorizontalProgressBarWithNunber extends ProgressBar {
         //当前进度和总值的比例
         float radio = getProgress() * 1.0f / getMax();
         //已到达的宽度
-        float progressPosX = (int) (mRealWidth * radio);
+        float progressPosX = mRealWidth * radio;
 
 
         // 绘制未到达的进度条
